Move lddmm modules to subpackage
