+++ herbrain/lddmm/__init__.py
@@ -1,0 +1,1 @@
+from ._lddmm import *
