+++ herbrain/lddmm/strings.py
